♻️ refactor(DetailsProjects): modularize project details component
- split DetailsProjects into smaller components for better readability and maintenance
- create HeroSection, ProjectDetails, TechnicalSidebar, and MediaGallery components
- move utility function getAssetImage to a separate utils file
