+++ src/components/DetailsProjects/utils.ts
@@ -1,0 +1,11 @@
+
+
+export const getAssetImage = (url: string): string => {
+  try {
+    const assetPath = url.replace("@/", "./");
+    return new URL(`/src/${assetPath}`, import.meta.url).href;
+  } catch (error) {
+    console.error("Error loading image:", error);
+    return "";
+  }
+};
